Rename `text2` to `text`
