+++ masonry/src/widget/textbox.rs
@@ -17,7 +17,7 @@ use vello::{
 use crate::widget::{LineBreaking, WidgetMut};
 use crate::{
     dpi::{LogicalPosition, LogicalSize},
-    text2::{TextBrush, TextEditor, TextStorage, TextWithSelection},
+    text::{TextBrush, TextEditor, TextStorage, TextWithSelection},
     AccessCtx, AccessEvent, BoxConstraints, CursorIcon, EventCtx, LayoutCtx, LifeCycle,
     LifeCycleCtx, PaintCtx, PointerEvent, StatusChange, TextEvent, Widget, WidgetId,
 };
